upgrade to glib-2.88.0

+++ gio/gmemoryinputstream.c
@@ -2,6 +2,8 @@
  * 
  * Copyright (C) 2006-2007 Red Hat, Inc.
  *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
@@ -30,16 +32,13 @@
 
 
 /**
- * SECTION:gmemoryinputstream
- * @short_description: Streaming input operations on memory chunks
- * @include: gio/gio.h
- * @see_also: #GMemoryOutputStream
+ * GMemoryInputStream:
  *
- * #GMemoryInputStream is a class for using arbitrary
+ * `GMemoryInputStream` is a class for using arbitrary
  * memory chunks as input for GIO streaming input operations.
  *
- * As of GLib 2.34, #GMemoryInputStream implements
- * #GPollableInputStream.
+ * As of GLib 2.34, `GMemoryInputStream` implements
+ * [iface@Gio.PollableInputStream].
  */
 
 struct _GMemoryInputStreamPrivate {
@@ -476,7 +475,7 @@ g_memory_input_stream_seek (GSeekable     *seekable,
       return FALSE;
     }
 
-  if (absolute < 0 || absolute > priv->len)
+  if (absolute < 0 || (gsize) absolute > priv->len)
     {
       g_set_error_literal (error,
                            G_IO_ERROR,
